+++ service/kas/kas.go
@@ -16,10 +16,10 @@ import (
 func NewRegistration() *serviceregistry.Service[kasconnect.AccessServiceHandler] {
 	return &serviceregistry.Service[kasconnect.AccessServiceHandler]{
 		ServiceOptions: serviceregistry.ServiceOptions[kasconnect.AccessServiceHandler]{
-			Namespace:      "kas",
-			ServiceDesc:    &kaspb.AccessService_ServiceDesc,
-			ConnectRPCFunc: kasconnect.NewAccessServiceHandler,
-			GRPCGateayFunc: kaspb.RegisterAccessServiceHandlerFromEndpoint,
+			Namespace:       "kas",
+			ServiceDesc:     &kaspb.AccessService_ServiceDesc,
+			ConnectRPCFunc:  kasconnect.NewAccessServiceHandler,
+			GRPCGatewayFunc: kaspb.RegisterAccessServiceHandler,
 			RegisterFunc: func(srp serviceregistry.RegistrationParams) (kasconnect.AccessServiceHandler, serviceregistry.HandlerServer) {
 				// FIXME msg="mismatched key access url" keyAccessURL=http://localhost:9000 kasURL=https://:9000
 				hostWithPort := srp.OTDF.HTTPServer.Addr