+++ service/wellknownconfiguration/wellknown_configuration.go
@@ -37,10 +37,10 @@ func RegisterConfiguration(namespace string, config any) error {
 func NewRegistration() *serviceregistry.Service[wellknownconfigurationconnect.WellKnownServiceHandler] {
 	return &serviceregistry.Service[wellknownconfigurationconnect.WellKnownServiceHandler]{
 		ServiceOptions: serviceregistry.ServiceOptions[wellknownconfigurationconnect.WellKnownServiceHandler]{
-			Namespace:      "wellknown",
-			ServiceDesc:    &wellknown.WellKnownService_ServiceDesc,
-			ConnectRPCFunc: wellknownconfigurationconnect.NewWellKnownServiceHandler,
-			GRPCGateayFunc: wellknown.RegisterWellKnownServiceHandlerFromEndpoint,
+			Namespace:       "wellknown",
+			ServiceDesc:     &wellknown.WellKnownService_ServiceDesc,
+			ConnectRPCFunc:  wellknownconfigurationconnect.NewWellKnownServiceHandler,
+			GRPCGatewayFunc: wellknown.RegisterWellKnownServiceHandler,
 			RegisterFunc: func(srp serviceregistry.RegistrationParams) (wellknownconfigurationconnect.WellKnownServiceHandler, serviceregistry.HandlerServer) {
 				wk := &WellKnownService{logger: srp.Logger}
 				return wk, nil